Fix Tests
Refactored the tests to import from base file.
